API review: move avocado.multiplexer to avocado.core.multiplexer
There's no current use case for direct use of the multiplexer APIs
from test code, so let's move that to avocado.core.
