+++ avocado/core/job.py
@@ -26,7 +26,6 @@ import tempfile
 import shutil
 import fnmatch
 
-from avocado import multiplexer
 from avocado import runtime
 from avocado import data_dir
 from avocado.core import runner
@@ -37,6 +36,7 @@ from avocado.core import exit_codes
 from avocado.core import exceptions
 from avocado.core import job_id
 from avocado.core import output
+from avocado.core import multiplexer
 from avocado.core.plugins import jsonresult
 from avocado.core.plugins import xunit
 from avocado.utils import archive
