+++ avocado/core/plugins/multiplexer.py
@@ -18,7 +18,7 @@ from avocado.core.plugins import plugin
 from avocado.core import output
 from avocado.core import exit_codes
 from avocado.core import tree
-from avocado import multiplexer
+from avocado.core import multiplexer
 
 
 class Multiplexer(plugin.Plugin):
