+++ avocado/core/plugins/runner.py
@@ -23,7 +23,7 @@ from avocado.core import exit_codes
 from avocado.core.plugins import plugin
 from avocado.core import output
 from avocado.core import job
-from avocado import multiplexer
+from avocado.core import multiplexer
 
 
 class TestRunner(plugin.Plugin):
